--- python/sdist/amici/conserved_moieties2.py
@@ -1,1 +1,0 @@
-../../amici/conserved_moieties2.py
